+++ Construct/Alphabet.cs
@@ -50,5 +50,7 @@ namespace PLGL.Construct
 
             return result.OrderBy(l => l.Key).ToList();
         }
+
+        
     }
 }